DML-320 NonNegativeInteger and NonNegativeIntegerDefault value moved a Double. Added Ocl constraint

+++ jsonmetaschemaMM/src-gen/jsonMetaschemaMM/impl/NonNegativeIntegerImpl.java
@@ -34,7 +34,7 @@ public class NonNegativeIntegerImpl extends MinimalEObjectImpl.Container impleme
 	 * @generated
 	 * @ordered
 	 */
-	protected static final int VALUE_EDEFAULT = 0;
+	protected static final double VALUE_EDEFAULT = 0.0;
 
 	/**
 	 * The cached value of the '{@link #getValue() <em>Value</em>}' attribute.
@@ -44,7 +44,7 @@ public class NonNegativeIntegerImpl extends MinimalEObjectImpl.Container impleme
 	 * @generated
 	 * @ordered
 	 */
-	protected int value = VALUE_EDEFAULT;
+	protected double value = VALUE_EDEFAULT;
 
 	/**
 	 * <!-- begin-user-doc -->
@@ -71,7 +71,7 @@ public class NonNegativeIntegerImpl extends MinimalEObjectImpl.Container impleme
 	 * @generated
 	 */
 	@Override
-	public int getValue() {
+	public double getValue() {
 		return value;
 	}
 
@@ -81,8 +81,8 @@ public class NonNegativeIntegerImpl extends MinimalEObjectImpl.Container impleme
 	 * @generated
 	 */
 	@Override
-	public void setValue(int newValue) {
-		int oldValue = value;
+	public void setValue(double newValue) {
+		double oldValue = value;
 		value = newValue;
 		if (eNotificationRequired())
 			eNotify(new ENotificationImpl(this, Notification.SET, JsonMetaschemaMMPackage.NON_NEGATIVE_INTEGER__VALUE, oldValue, value));
@@ -111,7 +111,7 @@ public class NonNegativeIntegerImpl extends MinimalEObjectImpl.Container impleme
 	public void eSet(int featureID, Object newValue) {
 		switch (featureID) {
 			case JsonMetaschemaMMPackage.NON_NEGATIVE_INTEGER__VALUE:
-				setValue((Integer)newValue);
+				setValue((Double)newValue);
 				return;
 		}
 		super.eSet(featureID, newValue);
